Airmed (Q1)/group-25-master/mysite/requirements.txt
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
colorama==0.4.1
control==0.8.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-chartit==0.2.9
django-chartjs==1.5.0
django-crispy-forms==1.7.2
entrypoints==0.3
gunicorn==19.9.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
jupyterlab==1.1.2
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.16.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
PyDSTool==0.90.2
pygame==1.9.4
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.0
PyQt5==5.12.1
PyQt5-sip==4.19.15
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pywinpty==0.5.5
pyzmq==18.1.0
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2